+++ CHANGELOG.md
@@ -1,3 +1,7 @@
-## v0.1.0 - August 24, 2022
+## v0.2.0 - August 24, 2022
+- **Breaking Change**: Change the reader and the parser initializer prototypes
+  - `cbor_reader_init()` and `cbor_parse()`
+  - Now a reader can be reusable to different messages
 
+## v0.1.0 - August 24, 2022
 The initial release.